concurrent-logging: make sure all log messages print
Use concurrently_ and forConcurrently_ from the async package instead
of forkIO.

Use a TVar to tell the printing thread when to stop instead of just
stopping when the queue is empty, because an empty queue doesn't
necessarily mean nothing else will get logged.

Use a continuation passing style to obtain the concurrent log action,
because this is encouraged by the async package to ensure that all
threads get canceled in the event of an exception.